pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "projet-2-modele-pour-demarrer"
version = "0.1.0"
[tool.poetry.dependencies]
Pillow = "^7.0"
flask = "^1.1"
python = "^3.8"
[tool.poetry.dev-dependencies]
